[review] 정규식 체크 함수를 String의 확장으로 만들지 않고, CommonString의 전역함수로 이동
